Oh buddy
Added A LOT of stuff and now want to go to sleep since its 6:30am..

-Added a Segmented control for Light/Dark map style in MapVC
-MapVC automatically detects if it’s night time and adjusts by turning
the map style to be dark, and light style when it’s day time. This is
also mimicked in the Navigation.
-Adjusted the Challenge + Logo Icon in MapVC to be lower in order to
support other screen sizes.
-Added FCAlertView Pod for a congratulations UIAlertController letting
the user know they completed a challenge, and how many points they
gained as well as the total points on the account.
-MapVC now only shows pins for the Root image instead of all images in
a group.
-ChallengeViewVC now shows all images in the same groupID.
-Tutorial screen now appropriately dismisses back to Profile if
navigated from the Profile, and navigates you to the Sign In screen if
it’s the first time loading the app.
-SettingsVC now contains rows to edit: Name, Username, Email. As well
as rows to select for Notification Settings (to be added today), and an
About screen with credits to us and OSU.
-FINALLY FIXED THE CAMERA CRASHING BUG. Turns out removing all of the
inputs and outputs is necessary when attempting to reuse the same
sessions/outlets/other stuff.
-Fixed AppLoadingVC to have the light status bar style.
-When a user now completes a Challenge, their Active Challenge is
reset/cleared back to empty so that they can start fresh.